Move Neptune to graphs/neptune folder
