+++ mem0/memory/main.py
@@ -139,7 +139,7 @@ class Memory(MemoryBase):
             if self.config.graph_store.provider == "memgraph":
                 from mem0.memory.memgraph_memory import MemoryGraph
             elif self.config.graph_store.provider == "neptune":
-                from mem0.memory.neptune.main import MemoryGraph
+                from mem0.graphs.neptune.main import MemoryGraph
             else:
                 from mem0.memory.graph_memory import MemoryGraph
